Hourly backup

+++ server.js
@@ -66,9 +66,10 @@ if(SECURE){
 	perMessageDeflate: false }).listen(PORT) })
 }else wss = new WebSocketServer({ port: PORT, perMessageDeflate: false })
 
-let criticalFiles = ["blacklist.txt", "../vip.txt", "webhook_url.text", "bansheets.txt", "config.json"]
+let criticalFiles = ["blacklist.txt", "../vip.txt", "webhook_url.text", "bansheets.txt", "./config.json"]
 for (let i = 0; i < criticalFiles.length; i++) {
-	await fs.writeFile(criticalFiles[i], "", err => { if (err) { console.error(err); return; } });
+	if (!await fsExists(criticalFiles[i])) await fs.writeFile(criticalFiles[i], "", err => { if (err) { console.error(err); return; } });
+
 }
 
 let players = 0
@@ -83,7 +84,7 @@ let hash = a => a.split("").reduce((a,b)=>(a*31+b.charCodeAt())>>>0,0)
 let allowed = new Set("rplace.tk google.com wikipedia.org pxls.space".split(" ")), censor = a => a.replace(/fuc?k|shi[t]|c[u]nt/gi,a=>"*".repeat(a.length)).replace(/https?:\/\/(\w+\.)+\w{2,15}(\/\S*)?|(\w+\.)+\w{2,15}\/\S*|(\w+\.)+(tk|ga|gg|gq|cf|ml|fun|xxx|webcam|sexy?|tube|cam|p[o]rn|adult|com|net|org|online|ru|co|info|link)/gi, a => allowed.has(a.replace(/^https?:\/\//,"").split("/")[0]) ? a : "").trim()
 
 wss.on('connection', async function(p, {headers, url: uri}) {
-	p.ip = headers['x-forwarded-for'].split(',').pop()
+	p.ip = headers['x-forwarded-for'].split(',').pop().split(':',4).join(':')
 	if(headers['origin'] != 'https://rplace.tk' || BANS.has(p.ip))return p.close()
 	let url = uri.slice(1)
 	let IP = /*p._socket.remoteAddress */url || p.ip
